+++ application/models/Estimates_model.php
@@ -988,6 +988,9 @@ class Estimates_model extends App_Model
             $this->db->where('rel_id', $estimate_id);
         }
         $this->db->where('rel_type', 'estimate');
+        if (function_exists('otmain_apply_files_not_deleted_filter')) {
+            otmain_apply_files_not_deleted_filter();
+        }
         $result = $this->db->get(db_prefix() . 'files');
         if (is_numeric($id)) {
             return $result->row();
@@ -1008,6 +1011,13 @@ class Estimates_model extends App_Model
         $attachment = $this->get_attachments('', $id);
         $deleted    = false;
         if ($attachment) {
+            if (function_exists('otmain_files_soft_delete_enabled')
+                && otmain_files_soft_delete_enabled()
+                && function_exists('otmain_soft_delete_file_row')
+            ) {
+                return otmain_soft_delete_file_row($attachment, 'Estimate');
+            }
+
             if (empty($attachment->external)) {
                 unlink(get_upload_path_by_type('estimate') . $attachment->rel_id . '/' . $attachment->file_name);
             }
