+++ application/models/Invoices_model.php
@@ -1142,6 +1142,9 @@ class Invoices_model extends App_Model
         }
 
         $this->db->where('rel_type', 'invoice');
+        if (function_exists('otmain_apply_files_not_deleted_filter')) {
+            otmain_apply_files_not_deleted_filter();
+        }
         $result = $this->db->get(db_prefix() . 'files');
         if (is_numeric($id)) {
             return $result->row();
@@ -1161,6 +1164,13 @@ class Invoices_model extends App_Model
         $attachment = $this->get_attachments('', $id);
         $deleted    = false;
         if ($attachment) {
+            if (function_exists('otmain_files_soft_delete_enabled')
+                && otmain_files_soft_delete_enabled()
+                && function_exists('otmain_soft_delete_file_row')
+            ) {
+                return otmain_soft_delete_file_row($attachment, 'Invoice');
+            }
+
             if (empty($attachment->external)) {
                 unlink(get_upload_path_by_type('invoice') . $attachment->rel_id . '/' . $attachment->file_name);
             }
